routerrpc: move payment intent extraction into package
