+++ lnrpc/routerrpc/config_active.go
@@ -3,7 +3,6 @@
 package routerrpc
 
 import (
-	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/lightningnetwork/lnd/macaroons"
 	"github.com/lightningnetwork/lnd/routing"
 )
@@ -22,12 +21,6 @@ type Config struct {
 	// NetworkDir is the main network directory wherein the router rpc
 	// server will find the macaroon named DefaultRouterMacFilename.
 	NetworkDir string
-
-	// ActiveNetParams are the network parameters of the primary network
-	// that the route is operating on. This is necessary so we can ensure
-	// that we receive payment requests that send to destinations on our
-	// network.
-	ActiveNetParams *chaincfg.Params
 
 	// MacService is the main macaroon service that we'll use to handle
 	// authentication for the Router rpc server.
